+++ src/onegov.feriennet/onegov/feriennet/views/settings.py
@@ -12,8 +12,13 @@ def settings_form(model, request):
 
     class CustomFieldsForm(Form):
         show_political_municipality = BooleanField(
-            label=_("Require the political municipality in the userprofile")
-        )
+            label=_("Require the political municipality in the userprofile"))
+
+        show_related_contacts = BooleanField(
+            label=_((
+                "Parents can see the contacts of other parents in "
+                "the same activity"
+            )))
 
         def process_obj(self, obj):
             super().process_obj(obj)
@@ -21,11 +26,17 @@ def settings_form(model, request):
             self.show_political_municipality.data = obj.meta.get(
                 'show_political_municipality', False)
 
+            self.show_related_contacts.data = obj.meta.get(
+                'show_related_contacts', False)
+
         def populate_obj(self, obj, *args, **kwargs):
             super().populate_obj(obj, *args, **kwargs)
 
             obj.meta['show_political_municipality']\
                 = self.show_political_municipality.data
+
+            obj.meta['show_related_contacts']\
+                = self.show_related_contacts.data
 
     return merge_forms(SettingsForm, CustomFieldsForm)