+++ src/onegov.feriennet/onegov/feriennet/models/__init__.py
@@ -1,4 +1,5 @@
 from onegov.feriennet.models.activity import VacationActivity
+from onegov.feriennet.models.calendar import Calendar, AttendeeCalendar
 from onegov.feriennet.models.invoice_action import InvoiceAction
 from onegov.feriennet.models.message import ActivityMessage
 from onegov.feriennet.models.message import PeriodMessage
@@ -6,6 +7,8 @@ from onegov.feriennet.models.notification_template import NotificationTemplate
 
 __all__ = [
     'ActivityMessage',
+    'AttendeeCalendar',
+    'Calendar',
     'InvoiceAction',
     'NotificationTemplate',
     'PeriodMessage',