+++ src/onegov.feriennet/HISTORY.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-- Adds a subscribable calendar to each attendee.
+- Adds a subscriptable calendar to each attendee.
   [href]
 
 - Optionally shows contact details of other parents for the same activity.